% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.24.01.22.R
\name{blast.FUNction}
\alias{blast.FUNction}
\title{Blast wrapper}
\usage{
blast.FUNction(
  query.fasta = "WADDL_DiagnosticPCR_Primers_dummy.fas",
  genome.fasta = "test_temp.fasta",
  word_size = 11,
  max_tar = 1
)
}
\arguments{
\item{query.fasta}{fasta file of sequence to search for}

\item{genome.fasta}{fasta file where the query.fasta sequences are searched for}
}
\value{
hit table
}
\description{
Run local blast and import the results as text file
}
