% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.24.01.22.R
\name{filter.outliers}
\alias{filter.outliers}
\title{Remove outliers}
\usage{
filter.outliers(var)
}
\arguments{
\item{var}{numeric vector}
}
\value{
numeric vector without outliers
}
\description{
Remove outliers from vector
}
