% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.24.01.22.R
\name{group.nmds}
\alias{group.nmds}
\title{Plots NMDS from OTU table}
\usage{
group.nmds(
  otu = otu.nasal,
  predictor = map.nasal$`MOVI PCR`,
  groups = c("positive", "negative"),
  size = 1,
  color = map.nasal$col1,
  return.table = F,
  spider = T,
  legend.location = "bottomleft"
)
}
\arguments{
\item{otu}{site abundance table}

\item{groups}{categories vector}
}
\value{
t test results
}
\description{
Plots NMDS from OTU table
}
