% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.24.01.22.R
\name{pairwiseRF}
\alias{pairwiseRF}
\title{makes figure illustrating taxa that are shared between two sites and unique to either site}
\usage{
pairwiseRF(
  mat = otu.fecal,
  map = map.fecal,
  dis_1 = "positive",
  dis_2 = "negative",
  num_of_col = 50,
  Accuray.cutoff = 0,
  Factor = 1,
  plot = T
)
}
\arguments{
\item{mat}{matrix or OTU table}

\item{map}{met data, need column named "discription"}
}
\value{
RF results
}
\description{
makes figure illustrating taxa that are shared between two sites and unique to either site
}
