% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.24.01.22.R
\name{plot.1var}
\alias{plot.1var}
\title{Plots 1 numeric vector against 1 categorical vector}
\usage{
\method{plot}{`1var`}(
  var = map.nasal$beta.disp,
  groups = map.nasal$discription.MOV2,
  col = map.nasal$col3,
  exclude.outliers = TRUE
)
}
\arguments{
\item{var}{numeric vector}

\item{groups}{categories vector}
}
\value{
t test results
}
\description{
Plots 1 numeric vector against 1 categorical vector
}
