% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.24.01.22.R
\name{remove_rare}
\alias{remove_rare}
\title{makes figure illustrating taxa that are shared between two sites and unique to either site}
\usage{
remove_rare(table, cutoff_pro)
}
\arguments{
\item{table}{site abundance table}

\item{cutoff_pro}{proportion of sites that an organism must be observed in ranges from 0 to 1}
}
\value{
filtered site abundance table
}
\description{
makes figure illustrating taxa that are shared between two sites and unique to either site
}
